gym_fullstack/backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
djongo>=1.3.6
pymongo[srv]>=4.6
django-cors-headers>=4.3
